Refactor rendering, core player state, and parallelize deployment pipeline.
This major update overhauls several critical systems across the workspace:

- Rendering & Visuals: Implemented a new nameplate rendering system with
  improved performance and dynamic sizing. Added spring-overshoot animations
  for world effects and status icons. Refactored building upgrade visualizations.
- Core Player Logic: Massive refactor of player state management in sow-core,
  improving how statistics (kills, troops, time alive) are tracked and serialized.
- HUD & UI: Extensive updates to the HUD, main menu, and lobby browser for
  better layout responsiveness and feature parity across platforms.
- Deployment Pipeline: Parallelized the sow-dist deployment into a four-phase
  pipeline (build, preparation, synchronization, and verification),
  significantly speeding up production and PTR releases.
- Tribe System: Introduced a historical tribe name generator and stable
  spirit animal emoji mapping for tribes.
- Tooling: Updated map migration, asset generation, and emoji atlas tools
  to support the latest rendering and gameplay changes.
- Balance & Config: Tuned game simulation parameters, including combat
  pacing, income rates, and building costs.

+++ sow-client/src/render/interact/actions.rs
@@ -19,8 +19,7 @@ impl SowApp {
                     self.sim.my_player_id = Some(1);
                     self.sim.my_lobby_id = Some(0);
 
-                    let map_id =
-                        sow_ui::ui::asset_loader::AssetLoader::map_key(&config.map_name);
+                    let map_id = sow_ui::ui::asset_loader::AssetLoader::map_key(&config.map_name);
                     self.ui.app.main_menu_state.downloading_map_name = Some(map_id.clone());
 
                     let mut config = *config;
@@ -38,9 +37,7 @@ impl SowApp {
                         &map_id,
                         crate::map_cache::load(&map_id),
                     ) {
-                        if let Ok(map_file) =
-                            sow_core::maps::load_map_from_payload(&payload)
-                        {
+                        if let Ok(map_file) = sow_core::maps::load_map_from_payload(&payload) {
                             config.map_width = map_file.width;
                             config.map_height = map_file.height;
                             self.ui
@@ -67,12 +64,7 @@ impl SowApp {
                                     format!("[{}] {}", tag, name)
                                 }
                             },
-                            color: self
-                                .ui
-                                .app
-                                .main_menu_state
-                                .selected_leader
-                                .filler_rgb(),
+                            color: self.ui.app.main_menu_state.selected_leader.filler_rgb(),
                             player_type: sow_core::player::PlayerType::Human,
                             team: None,
                             spawn_x: 0,
@@ -226,8 +218,8 @@ impl SowApp {
                     }
                 }
                 UiAction::FocusTile(col, row) => {
-                    let world_cx = col as f32 + 0.5;
-                    let world_cy = row as f32 + 0.5;
+                    let world_cx = col + 0.5;
+                    let world_cy = row + 0.5;
 
                     // Zoom in to a comfortable battle-focus level
                     let target_zoom = 3.0_f32.max(self.input.camera_zoom);
@@ -268,11 +260,14 @@ impl SowApp {
                     );
                 }
                 UiAction::CopyInviteLink(lobby_id) => {
-                    if let Some(link) = crate::store_portals::invite_link(lobby_id, &crate::get_build_version()) {
+                    if let Some(link) =
+                        crate::store_portals::invite_link(lobby_id, &crate::get_build_version())
+                    {
                         self.ui.egui_ctx.copy_text(link);
                         // Approximate time using instant since we don't have a direct timer in this context
                         // or we can just use the ui input time.
-                        self.ui.app.main_menu_state.invite_copied_at = Some(self.ui.egui_ctx.input(|i| i.time));
+                        self.ui.app.main_menu_state.invite_copied_at =
+                            Some(self.ui.egui_ctx.input(|i| i.time));
                         log::info!("Copied invite link to clipboard.");
                     }
                 }
